requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
nvidia-ml-py==12.575.51
openai-whisper==20240930
packaging==25.0
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pynvml==12.0.0
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
